Fixed WebSocket bugs and Updated README.md

+++ src/main/java/edu/kit/tm/cm/scdm/sensordatastorage/application/dtos/response/CoordinateResponse.java
@@ -14,7 +14,7 @@ public class CoordinateResponse {
     private double latitude;
 
     @ApiModelProperty(position = 2, required = true)
-    private float longitude;
+    private double longitude;
